split join_exports.cpp
